requirements.txt
Django==1.7
Fabric==1.10.0
Pillow==2.6.1
Unidecode==0.04.16
argparse==1.2.1
args==0.1.0
beautifulsoup4==4.3.2
configparser==3.3.0r2
django-app-namespace-template-loader==0.2
django-appconf==0.6
django-blog-zinnia==0.15
django-braces==1.4.0
django-ckeditor==4.4.7
django-compressor==1.4
django-configurations==0.8
django-constance==1.0
django-contrib-comments==1.5
django-crispy-forms==1.4.0
django-extensions==1.4.0
django-filter==0.8
django-filters-mixin==0.1.2
django-floppyforms==1.3.0
django-modeltranslation==0.8
django-mptt==0.6.1
django-picklefield==0.3.1
django-recaptcha==1.0.2
django-tables2==0.15.0
django-tagging==0.3.4
django-tinymce==1.5.3
django-xmlrpc==0.1.5
ecdsa==0.11
fabtools==0.19.0
paramiko==1.15.1
psycopg2==2.5.4
pycrypto==2.6.1
pyparsing==2.0.3
pytz==2014.10
raven==5.0.0
redis==2.10.3
six==1.8.0
wsgiref==0.1.2
-e git+https://github.com/Fantomas42/zinnia-theme-bootstrap.git@526342a2fb32ddf1e5eb2a9f13869992eb6f3d23#egg=zinnia_theme_bootstrap-dev
zinnia-wysiwyg-ckeditor==1.2
